+++ src/mocks/handlers.js
@@ -8,8 +8,7 @@ import {
   getFestivalReviewHandler,
   getRestaurantReviewHandler,
   getReviewedHandler,
-  postFestivalReviewHandler,
-  postRestaurantReviewHandler,
+  getTouristSpotReviewHandler,
 } from "./get/review";
 import { getTouristSpotHandler } from "./get/touristSpot";
 import { searchFoodHandler, getFoodHandler } from "./get/food";
@@ -24,6 +23,11 @@ import {
   postFestivalReservationHandler,
   postRestaurantReservationHandler,
 } from "./post/reservation";
+import {
+  postFestivalReviewHandler,
+  postRestaurantReviewHandler,
+  postTouristSpotReviewHandler,
+} from "./post/review";
 
 export const handlers = [
   searchSpotHandler,
@@ -35,6 +39,7 @@ export const handlers = [
   getTouristSpotHandler,
   getRestaurantReviewHandler,
   getFestivalReviewHandler,
+  getTouristSpotReviewHandler,
   searchFoodHandler,
   getFoodHandler,
   getRestaurantReservationHandler,
@@ -45,9 +50,10 @@ export const handlers = [
   registerHandler,
   getUserinfoHandler,
   editUserHandler,
-  postFestivalReviewHandler,
-  postRestaurantReviewHandler,
   postRestaurantReservationHandler,
   postFestivalReservationHandler,
   getReviewedHandler,
+  postRestaurantReviewHandler,
+  postFestivalReviewHandler,
+  postTouristSpotReviewHandler,
 ];
